refactor semiring stuff
