+++ examples/marginal_map_experiment.rs
@@ -7,7 +7,8 @@ use rand::Rng;
 use rsdd::{
     builder::{bdd_builder::BddManager, cache::all_app::AllTable},
     repr::{cnf::Cnf, robdd::BddPtr, var_label::VarLabel, wmc::WmcParams},
-    util::semiring::{RealSemiring, Semiring},
+    util::semirings::semiring_traits::Semiring,
+    util::semirings::realsemiring::RealSemiring
 };
 
 #[derive(Parser, Debug)]
