+++ src/builder/bdd_builder.rs
@@ -655,7 +655,8 @@ mod tests {
 
     use crate::builder::BottomUpBuilder;
     use crate::repr::wmc::WmcParams;
-    use crate::util::semiring::{RealSemiring, Semiring};
+    use crate::util::semirings::realsemiring::RealSemiring;
+    use crate::util::semirings::semiring_traits::Semiring;
     use crate::{builder::cache::all_app::AllTable, repr::ddnnf::DDNNFPtr};
     use maplit::*;
