+++ src/repr/sdd/binary_sdd.rs
@@ -10,7 +10,7 @@ use crate::{
         vtree::{VTreeIndex, VTreeManager},
         wmc::WmcParams,
     },
-    util::semiring::FiniteField,
+    util::semirings::finitefield::FiniteField,
 };
 
 use super::SddPtr;
